untested: arm_mpam: resctrl: Add support for MB resource
ANBZ: #8686

commit 62e0c4c1568edc593aeb130a3f83e628999df009 morse-linux.

resctrl supports 'MB', as a percentage throttling of traffic somewhere
after the L3. This is the control that mba_sc uses, so ideally the
class chosen should be as close as possible to the counters used for
mba_local.

MB's percentage control can be backed either with the fixed point
fraction MBW_MAX or the bandwidth portion bitmap. Add helper to convert
to/from percentages. One problem here is the value written is not the
same as the value read back. This is deliberatly made visible to
user-space. Another is the MBW_MAX fixed point fraction can't represent
100%. This is also exposed to user-space, as otherwise the values for a
single-bit system is 100%, 0%, instead of 50%, 0%.

The way CDP is emulated means MB controls need programming twice by
the resctrl glue, as the bandwidth controls can be applied independently
for data or instruction-fetch. This isn't how x86 behaves, and neither
user-space nor resctrl support it.

CC: Amit Singh Tomar <amitsinght@marvell.com>
Signed-off-by: James Morse <james.morse@arm.com>
[ cherry-picked from https://git.kernel.org/pub/scm/linux/kernel/git/morse/linux.git/log/?h=mpam/snapshot/v6.7-rc2 ]
Signed-off-by: Shawn Wang <shawnwang@linux.alibaba.com>
Reviewed-by: Baolin Wang <baolin.wang@linux.alibaba.com>
Link: https://gitee.com/anolis/cloud-kernel/pulls/3011