net: hns3: add support PF provides customized interfaces to detect port faults.
ANBZ: #19067

commit 13dc8adbfa8fc704da0824788dc4236d9ea67ef0 openEuler.

driver inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I97T47
CVE: NA

----------------------------------------------------------------------

When the link status of the PF network port changes from up to down,
the IMP can quickly report the link down fault cause based on the
link status. If the hardware is faulty from the beginning and the
link is not up, the IMP does not report the link down event
because the status is not changed.

Therefore, an interface is provided to detect port faults.

Signed-off-by: Jie Wang <wangjie125@huawei.com>
Signed-off-by: Jiantao Xiao <xiaojiantao1@h-partners.com>
Signed-off-by: wangkaiyuan <wangkaiyuan@inspur.com>
Reviewed-by: Bitao Hu <yaoma@linux.alibaba.com>
Reviewed-by: Guixin Liu <kanie@linux.alibaba.com>
Link: https://gitee.com/anolis/cloud-kernel/pulls/4834