anolis: arm64: support splitting the block mapping kernel page table
ANBZ: #8540

When the linear mapping range adopts block or contiguous mapping,
the kernel can no longer support to modify the attribute of kernel
page table in PTE granularity. Scenarios like RODATA/KFENCE/Crash
kernel/Memory failure needs the capability to modify the page table
attribute in PTE granularity  to ensure the correctness and
reliability. However, if the linear mapping range is mapped by PTE
level, the ratio of TLB miss would increase obviously and the
performance of programs would decrease when accessing the linear
mapping areas. Currently, the performance and the attribute control
of PTE level are incompatible. In practice, changing the page table
attributes of PTE granularity is a low-frequency and occasional
behavior, which gives us the opportunity to selectively split the
block/contiguous mapping into PTE granularity to achieve both
requirements above.

This patch provides splitting the block/contiguous mapping dynamicly.
To avoid the behavior of TLB prefetch, the process of splitting
follows break-before-make principle to avoid TLB conflict. Meanwhile,
if cleared the kernel page table entry, it is necessary to avoid
accessing corresponding address that might be cleared before. In the
system boot stage, only one CPU is working. There is no need to
consider other CPUs access the cleared range simultaneously.

In addition, when initializing the kernel page table, we should avoid
the situation that the physical range that the cleared page table
entry pointed to contains the physical address of the page table entry.
It would cause a kernel page fault if dynamically splitting the page
table entry later.

Signed-off-by: Kaihao Bai <carlo.bai@linux.alibaba.com>
Reviewed-by: Baolin Wang <baolin.wang@linux.alibaba.com>
Link: https://gitee.com/anolis/cloud-kernel/pulls/2885