net: hns3: Resolved the issue that the debugfs query result is inconsistent.
ANBZ: #19067

commit 233f4df4386484caa52a3ed036f8e983795e0c0c openEuler.

driver inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/IASZI4
CVE: NA

----------------------------------------------------------------------

This patch modifies the implementation of debugfs:
When the user process stops unexpectedly, not all data of the file system
is read. In this case, the save_buf pointer is not released. When the user
process is called next time, save_buf is used to copy the cached data
to the user space. As a result, the queried data is inconsistent. To solve
this problem, determine whether the function is invoked for the first time
based on the value of *ppos. If *ppos is 0, obtain the actual data.

Fixes: c0d3248b8d92 ("net: hns3: refactor the debugfs process")
Signed-off-by: Guangwei Zhang <zhangwangwei6@huawei.com>
Signed-off-by: Hao Lan <lanhao@huawei.com>
Signed-off-by: Hao Chen <chenhao418@huawei.com>
Signed-off-by: wangkaiyuan <wangkaiyuan@inspur.com>
Reviewed-by: Bitao Hu <yaoma@linux.alibaba.com>
Reviewed-by: Guixin Liu <kanie@linux.alibaba.com>
Link: https://gitee.com/anolis/cloud-kernel/pulls/4873