anolis: configs: generate config files with dummy pahole
ANBZ: #9853

Some kconfigs are enable/disabled depends on the toolchain's version.
Most of these differences could be avoid by using dummy gcc, ld
and nm located in scripts/dummy-tools/, with cmd `make
CROSS_COMPILE=scripts/dummy-tools/ xxxconfig`, but CONFIG_PAHOLE_VERSION
is an exception, it still depends on the host's pahole tool.

If pahole is not installed on host, CONFIG_MODULE_ALLOW_BTF_MISMATCH
will be invisible in the final generated config files. And then if
`make listnewconfig` is checked on another host with pahole installed,
it will warning that this config is newly imported.

To fix this problem, generating config files with dummy pahole tool.

Fixes: 924c80eb23ee ("anolis: configs: add kconfigs baseline framework")
Signed-off-by: Qiao Ma <mqaio@linux.alibaba.com>
Link: https://gitee.com/anolis/cloud-kernel/pulls/3793