anolis: crypto: ccp: Introduce hygon specific interface to support driver
ANBZ: #11419

Hygon secure processors provide a lot of security functions, which
require a lot of code to support. In order to prevent Hygon function
code from invading the driver's native code, we introduce specific
files for Hygon. We'll leave the native code unchanged as much as
possible.

In this patch, we add files as below:

    a. files for codes to support Hygon secure processor:
        drivers/crypto/ccp/hygon/sp-dev.h
        drivers/crypto/ccp/hygon/sp-pci.c
        drivers/crypto/ccp/hygon/psp-dev.c
        drivers/crypto/ccp/hygon/psp-dev.h

    b. header file to define data types and structures for HYGON
       Platform Security Processor:
        include/linux/psp-hygon.h

    c. header file to define userspace interface for HYGON Platform
       Security Processor:
        include/uapi/linux/psp-hygon.h

We'll add more Hygon specific code in the following commits.

In the following commits, we'll move the code for Hygon from the native
file to the Hygon specific file.

Signed-off-by: hanliyang <hanliyang@hygon.cn>
Reviewed-by: Xingrui Yi <yixingrui@linux.alibaba.com>
Reviewed-by: Artie Ding <artie.ding@linux.alibaba.com>
Link: https://gitee.com/anolis/cloud-kernel/pulls/4009