anolis: mm: add filemap_range_is_marked() helper
ANBZ: #18906

Sometimes we want to check whether a specific tag exists within a
certain range of the address space. For example, in the FUSE file
system, we need to check whether the page cache of a file contains a
specific tag in order to perform the corresponding logic.

This patch adds the filemap_range_is_marked() helper to accomplish
this.

Signed-off-by: Hongzhen Luo <hongzhen@linux.alibaba.com>
Reviewed-by: Jingbo Xu <jefflexu@linux.alibaba.com>
Reviewed-by: Joseph Qi <joseph.qi@linux.alibaba.com>
Reviewed-by: Baolin Wang <baolin.wang@linux.alibaba.com>
Link: https://gitee.com/anolis/cloud-kernel/pulls/4643