anolis: KVM: arm64: Flush TLB during module loading
ANBZ: #4924

When the bitmap is empty, it assumes that the TLB is clean and all VMIDs
are unused (see flush_context()). However, when KVM is built as a
module, there may be leftover dirty TLB entries from the previous module
loading. Therefore, the TLB should be flushed here to ensure that there
are no dirty TLB entries remaining.

Fixes: 9932aa87b9b1 ("anolis: KVM: arm64: Adapt KVM to be built as a module")
Signed-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>
Reviewed-by: Wu Zongyong <wuzongyong@linux.alibaba.com>
Reviewed-by: Shuai Xue <xueshuai@linux.alibaba.com>
Link: https://gitee.com/anolis/cloud-kernel/pulls/4720